pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forscheule"
version = "0.1.0"
description = "Daily paper suggestions for spatial transcriptomics research"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31,<3",
    "python-dotenv>=1.0,<2",
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27,<1",
    "torch>=2.1",
    "transformers>=4.36",
    "rapidfuzz>=3.6,<4",
    "numpy>=1.26,<3",
    "jinja2>=3.1,<4",
    "openai>=1.30,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "ruff>=0.3,<1",
    "httpx>=0.27,<1",
]

[tool.setuptools.packages.find]
include = ["forscheule*"]

[tool.setuptools.package-data]
forscheule = ["templates/*.html"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
